+++ packages/mobile/src/App.tsx
@@ -48,6 +48,7 @@ import { DuplicatedUsernameScreen } from './screens/DuplicatedUsername/Duplicate
 import { UnregisteredUsernameScreen } from './screens/UnregisteredUsername/UnregisteredUsername.screen'
 import UsernameTakenScreen from './screens/UsernameTaken/UsernameTaken.screen'
 import NewUsernameRequestedScreen from './screens/NewUsernameRequested/NewUsernameRequested.screen'
+import { PossibleImpersonationAttackScreen } from './screens/PossibleImpersonationAttack/PossibleImpersonationAttack.screen'
 
 LogBox.ignoreAllLogs()
 
@@ -107,6 +108,10 @@ function App(): JSX.Element {
                 <Screen component={UnregisteredUsernameScreen} name={ScreenNames.UnregisteredUsernameScreen} />
                 <Screen component={UsernameTakenScreen} name={ScreenNames.UsernameTakenScreen} />
                 <Screen component={NewUsernameRequestedScreen} name={ScreenNames.NewUsernameRequestedScreen} />
+                <Screen
+                  component={PossibleImpersonationAttackScreen}
+                  name={ScreenNames.PossibleImpersonationAttackScreen}
+                />
               </Navigator>
               <CommunityContextMenu />
               <ChannelContextMenu />