+++ packages/mobile/src/store/navigation/redirection/redirection.saga.ts
@@ -3,7 +3,7 @@ import { initSelectors } from '../../init/init.selectors'
 import { navigationSelectors } from '../navigation.selectors'
 import { navigationActions } from '../navigation.slice'
 import { ScreenNames } from '../../../const/ScreenNames.enum'
-import { identity, publicChannels } from '@quiet/state-manager'
+import { identity, publicChannels, users } from '@quiet/state-manager'
 import { initActions } from '../../init/init.slice'
 
 export function* redirectionSaga(): Generator {
@@ -31,6 +31,16 @@ export function* redirectionSaga(): Generator {
     yield* put(
       navigationActions.replaceScreen({
         screen: ScreenNames.UsernameTakenScreen,
+      })
+    )
+  }
+
+  const duplicateCerts = yield* select(users.selectors.duplicateCerts)
+
+  if (duplicateCerts) {
+    yield* put(
+      navigationActions.replaceScreen({
+        screen: ScreenNames.PossibleImpersonationAttackScreen,
       })
     )
     return