+++ packages/state-manager/src/index.ts
@@ -82,8 +82,6 @@ export * from './constants'
 
 export { formatBytes } from './utils/functions/formatBytes/formatBytes'
 
-export { sortPeers } from './utils/functions/sortPeers/sortPeers'
-
 export { getInvitationCodes } from './utils/functions/invitationCode/invitationCode'
 
 export type { Socket } from './types'