+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+
 [unreleased]
 
 * UI layer for taken usernames for desktop and mobile
@@ -8,6 +9,8 @@
 
 * Update registrar service to match new registration flow.
 
+* Add possible impersonation attack UI for desktop and mobile
+
 * Fix truncated long messages in channelInput component
 
 * Unblock mobile e2e tests
@@ -15,6 +18,8 @@
 * Prettify loading component on Chat screen (mobile)
 
 * Running Chromatic tests for forked PRs
+
+* Added e2e test for user joining community when owner is offline. Improved e2e tests
 
 * Bump github actions/* to versions using node16